Ignore non-file URIs, to avoid validating eg. diff preview windows
Ran into this when testing out the test suggestions quickfix stuff- if I
previewed the changes, the CLI would conk out and I'd see a log message
about test names needing to be unique. This is because the preview
window got picked up by the LSP and run through the validator. Whoops!